requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
discord-webhook==1.3.1
dnspython==2.6.1
idna==3.7
pymongo==4.8.0
python-dotenv==1.0.1
redis==5.0.7
requests==2.32.3
spotipy==2.24.0
urllib3==2.2.2
